+++ src/bodyModule/body_moon.cpp
@@ -196,7 +196,7 @@ void Moon::drawBody(const Projector* prj, const Navigator * nav, const Mat4d& ma
 	//tesselation
 	if ( myShader == SHADER_MOON_NORMAL_TES) {
 		myShaderProg->setUniform("TesParam", 
-				Vec3f(bodyTesselation->min_tes_level,bodyTesselation->max_tes_level, bodyTesselation->moon_altimetry_factor) );
+				Vec3i(bodyTesselation->min_tes_level,bodyTesselation->max_tes_level, bodyTesselation->moon_altimetry_factor) );
 	}
 
 	if ( myShader == SHADER_MOON_NORMAL_TES )
